pyproject.toml
[project]
name = "mcpservercleinthost"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "firebase-admin>=7.4.0",
    "mcp>=1.27.0",
    "openai>=2.32.0",
    "pydantic>=2.13.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
